Update to the curation SOP
Page 1: Updated date and version of the SOP. Updated to July 2021 Version 1.1
Page 7: Defined clinician and clarified qualifications of experts who contribute clinical expertise. Page 7 now states “Clinicians are those with a clinical doctorate, including but not limited to those with an MD, DO, or PharmD, and will contribute clinical expertise.”
Page 11: Updated to correctly state the authoritative resources as indicated in page 10. Page 11 has been updated to correctly state “ClinVar’s 3-star expert panel and 4-star practice guideline submissions”
Page 15: Clarified qualifications of clinicians. Added “e.g.” in the parentheticals before “MD or PharmD” to reflect that these degrees are examples of degrees held by clinicians and is not an inclusive list.